LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/resource-timing/object-not-found-after-TAO-cross-origin-redirect-expected.txt


Harness Error (TIMEOUT), message = null

PASS window.performance is defined
PASS window.performance.getEntriesByName is defined
